+++ bubbles/service_utils.py
@@ -1,7 +1,20 @@
 import subprocess
 import time
 
-from bubbles import PROCESS_CHECK_SLEEP_TIME, PROCESS_CHECK_COUNT, get_service_name
+
+SERVICES = ["tor", "tor_ocr", "tor_archivist", "blossom", "all", "buttercup"]
+# special cases
+SERVICE_NAMES = {"tor": "tor_moderator"}
+PROCESS_CHECK_SLEEP_TIME = 10  # seconds
+PROCESS_CHECK_COUNT = 5
+
+
+def get_service_name(service: str) -> str:
+    # sometimes the services have different names on the server than we
+    # know them by. This translation layer helps keep track of that.
+    if service in SERVICE_NAMES:
+        return SERVICE_NAMES[service]
+    return service
 
 
 def say_code(say, message: bytes) -> None:
